Complete new method, working, weird results
This is the code I mentioned in the Christmas plan with more time saving methods and the new direct p(A|D,n). It works and calculates the peak but there are the following problems: 
The results are weird. I managed to get the factor almost to one but for most it oscillates from 0,2 to 8 or something like that. 
The shifted gaussians (for each point) should be higher for lower A and broader and lower for higher A if I remember correctly. This is not the case as they stay the same height but change the width... this is possible source of error.
Also all the probabilities stays lower than 1, although not being normalized. This means that the logsum is very negative number even though the shape of the thing looks okay (but there is no peak, I look at the logsum directly, not the exponential of it) - might be not a problem in the end.